agentic_rag/requirements.txt
docling
langchain-openai
chromadb
python-dotenv
openai
pydantic
fastapi
uvicorn
python-multipart
transformers
torch
accelerate